// React/app/GerAdv_TS/Graph/Interfaces/interface.Graph.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface IGraph {
// 202501251
    id: number;
 
	tabela: string,
	tabelaid: number,
	imagem: Uint8Array,
}

export interface IGraphFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IGraphIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}